Refactor types into dedicated module
The someip types should not be located in the parser module because
they're generally used by the other modules as well and they are not
parsing specific.
